Changed ship rotatio to feel more natural

+++ src/org/alien8/managers/ModelManager.java
@@ -21,7 +21,7 @@ import org.alien8.ship.Ship;
  */
 public class ModelManager {
 
-  private int lastSerial = 0;
+  private long lastSerial = 0;
   private static ModelManager instance = new ModelManager();
   private ConcurrentLinkedQueue<Entity> entities = new ConcurrentLinkedQueue<Entity>();
   private CollisionDetector collisionDetector = new CollisionDetector();
